code refactoring

+++ services/UserService.py
@@ -1,4 +1,8 @@
-from flask import Response
+from datetime import datetime
+
+from flask import Response, session
+from werkzeug.security import check_password_hash
+
 from dao.UserDAO import UserDao
 from models.User import User
 from utils.Convert import Convert
@@ -13,7 +17,7 @@ class UserService:
         self.validate = Validate()
 
     def get(self, id: int) -> dict | Response:
-        user = self.userDAO.get_by_id(id)
+        user = self.userDAO.get_by_id(id, User)
         if self.validate.check_user(user):
             try:
                 return self.convert.convert_data_user(user)
@@ -22,34 +26,66 @@ class UserService:
                 return Response(status=500)
         return Response(status=400)
 
+    def get_all(self) -> list:
+        data = []
+        for user in self.userDAO.get_all(User):
+            data.append(self.convert.convert_data_user(user))
+        return data
+
     def create(self, data: dict) -> dict | Response:
         if self.validate.check_data(data):
             try:
-                user = User(name=data.get('name'), email=data.get('email'))
-                return self.convert.convert_data_user(self.userDAO.create(user))
+                user = User(name=data.get('name'), email=data.get('email'), password=data.get('password'))
+                us = self.userDAO.create(user)
+                session['user_id'] = us.get_id()
+                return self.convert.convert_data_user(us)
             except Exception as err:
                 print(err)
                 return Response(status=500)
         return Response(status=400)
 
-    def update(self, id: int, _data: dict) -> User | Response:
-        user = self.userDAO.get_by_id(id)
-        try:
-            if self.validate.check_user(user):
-                data = self.validate.update_data(_data, user)
-                return self.convert.convert_data_user(self.userDAO.update(id, data))
-        except Exception as err:
-            print(err)
-            return Response(status=500)
-        return Response(status=400)
-
-    def delete(self, id: int) -> Response:
-        user = self.userDAO.get_by_id(id)
-        if self.validate.check_user(user):
+    def update(self, _data: dict) -> User | Response:
+        user_id = session.get('user_id')
+        if user_id:
+            user = self.userDAO.get_by_id(user_id, User)
             try:
-                self.userDAO.delete(user)
-                return Response(status=204)
+                if self.validate.check_user(user):
+                    data = self.validate.update_data(_data, user)
+                    return self.convert.convert_data_user(self.userDAO.update(user_id, data))
             except Exception as err:
                 print(err)
                 return Response(status=500)
-        return Response(status=404)
+            return Response(status=400)
+        return Response(status=401)
+
+    def delete(self) -> Response:
+        user_id = session.get('user_id')
+        if user_id:
+            user = self.userDAO.get_by_id(user_id, User)
+            if self.validate.check_user(user):
+                try:
+                    self.userDAO.delete(user)
+                    session.pop('user_id')
+                    return Response(status=204)
+                except Exception as err:
+                    print(err)
+                    return Response(status=500)
+            return Response(status=404)
+        return Response(status=401)
+
+    def login(self, data: dict) -> dict | Response:
+        email = data.get('email')
+        password = data.get('password')
+        user = self.userDAO.get_by_email(email)
+
+        if user and check_password_hash(user.password, password):
+            user.set_last_login(datetime.utcnow())
+            session['user_id'] = user.get_id()
+            return self.convert.convert_data_user(user)
+
+        return Response(status=400)
+
+    @staticmethod
+    def logout() -> Response:
+        session.pop('user_id')
+        return Response(status=200)
